requirements.txt
pandas~=1.3.5
sentence-transformers~=2.1.0
scikit-learn~=1.0.2
scipy~=1.7.1
pytorch~=1.9.1
numpy~=1.20.3
transformers~=4.11.3
easynmt~=2.1.0
